Silence various other compiler warnings
git-svn-id: svn+ssh://svn.code.sf.net/p/cmusphinx/code/trunk/pocketsphinx@6699 94700074-3cef-4d97-a70e-9c8c206c02f5

+++ src/libpocketsphinx/list.c
@@ -111,7 +111,7 @@ list_lookup(list_t const *list, int32 idx)
     static char const *rname = "list_lookup";
 
     if ((list == 0) || (idx >= list->size) || (idx < 0))
-        return ((caddr_t) exception(rname, "idx", ERR_ARG));
+        return ((caddr_t)(long)exception(rname, "idx", ERR_ARG));
 
     return (list->list[idx]);
 }
